EUI (excellent user interface)
Brought the same styling from the sign up page to the login page.

Deleted the Birthday field in the sign up page. We don’t need to know
how old booksellers are. We want to limit the amount of impatience a
user will experience when they’re led to sign up page.
On a side note:
I personally don’t think we need a bio of the user, maybe just make it
optional?