Form-Verification
In create_post.html, I had to change the submit button to be a button
and not a link because it was causing some issues in terms of checking
data.

In create_post.js, we’ve added a better form-verification snippet, and
I did the same thing before without putting it into an array but it
works now that we’ve put it into an array. Go figure